Move get_region and set_region functions into aws submodule. Update references throughout
