+++ cloudknot/__init__.py
@@ -4,8 +4,9 @@ import logging
 
 from . import aws  # noqa
 from . import config  # noqa
-from .dockerimage import *  # noqa
+from .aws.base_classes import get_region, set_region  # noqa
 from .cloudknot import *  # noqa
+from .dockerimage import *  # noqa
 from .version import __version__  # noqa
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
