+++ cloudknot/tests/test_aws.py
@@ -115,7 +115,7 @@ def test_ObjectWithUsernameAndMemory():
 
 
 def test_IamRole():
-    iam = boto3.client('iam', region_name=ck.config.get_region())
+    iam = boto3.client('iam', region_name=ck.get_region())
     config = configparser.ConfigParser()
     config_file = ck.config.get_config_file()
 
@@ -316,7 +316,7 @@ def test_IamRole():
 
 
 def test_DockerRepo():
-    ecr = boto3.client('ecr', region_name=ck.config.get_region())
+    ecr = boto3.client('ecr', region_name=ck.get_region())
     config = configparser.ConfigParser()
     config_file = ck.config.get_config_file()
 
@@ -426,7 +426,7 @@ def test_DockerRepo():
 
 
 def test_Vpc():
-    ec2 = boto3.client('ec2', region_name=ck.config.get_region())
+    ec2 = boto3.client('ec2', region_name=ck.get_region())
     config = configparser.ConfigParser()
     config_file = ck.config.get_config_file()
 
@@ -651,7 +651,7 @@ def test_Vpc():
 
 
 def test_SecurityGroup():
-    ec2 = boto3.client('ec2', region_name=ck.config.get_region())
+    ec2 = boto3.client('ec2', region_name=ck.get_region())
     config = configparser.ConfigParser()
     config_file = ck.config.get_config_file()
 
@@ -852,7 +852,7 @@ def test_SecurityGroup():
 
 
 def test_JobDefinition(pars):
-    batch = boto3.client('batch', region_name=ck.config.get_region())
+    batch = boto3.client('batch', region_name=ck.get_region())
     config = configparser.ConfigParser()
     config_file = ck.config.get_config_file()
 
@@ -1090,7 +1090,7 @@ def test_JobDefinition(pars):
 
 
 def test_ComputeEnvironment(pars):
-    batch = boto3.client('batch', region_name=ck.config.get_region())
+    batch = boto3.client('batch', region_name=ck.get_region())
     config = configparser.ConfigParser()
     config_file = ck.config.get_config_file()
 
@@ -1592,7 +1592,7 @@ def test_ComputeEnvironment(pars):
 
 
 def test_JobQueue(pars):
-    batch = boto3.client('batch', region_name=ck.config.get_region())
+    batch = boto3.client('batch', region_name=ck.get_region())
     config = configparser.ConfigParser()
     config_file = ck.config.get_config_file()
